requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.9
psycopg2-binary==2.9.9

